debug: sleep forever on the init process
This change introduces a change to sleep forever, right after runc
creates the init process. By doing this we delay the next steps of the
runc exectuion flow, allowing us to debug issues.

The issues we are interested here are the ones that happens in a
Kubernetes cluster where runc is the underlying container shim used by
containerd to create the containers specific to the Pods. We are
interested in scenarios - where within the small time gap between when the
runc init process comes up and it execs into the actual container
process - something happens like a node shutdown or disconnection etc.
This puts the runc init process at a hanging state.

We want to be able reproduce this and work towards organic means to
cleanup the erroneous state.